API reference

Index

PseudoPotentialIO.HghFile

PseudoPotentialIO.PsPFile

PseudoPotentialIO.Psp8File

PseudoPotentialIO.UpfFile

PseudoPotentialIO.UpfFile

PseudoPotentialIO.element

PseudoPotentialIO.formalism

PseudoPotentialIO.format

PseudoPotentialIO.functional

PseudoPotentialIO.has_model_core_charge_density

PseudoPotentialIO.has_spin_orbit

PseudoPotentialIO.identifier

PseudoPotentialIO.is_norm_conserving

PseudoPotentialIO.is_paw

PseudoPotentialIO.is_ultrasoft

PseudoPotentialIO.load_psp_file

PseudoPotentialIO.save_psp_file

PseudoPotentialIO.valence_charge

Types

PseudoPotentialIO.PsPFile — Type

abstract type PsPFile

Abstract type representing a pseudopotential file.

The structure of the data should closely mirror the format of the file, and the values of quantities should be exactly those found in the file.

PseudoPotentialIO.UpfFile — Type

struct UpfFile <: PsPFile

Universal Pseudopotential Format file contents.

identifier::String: Identifier

version::String: UPF format version

info::Union{Nothing, String}: Optional general information about the pseudopotential, often generation input

header::PseudoPotentialIO.UpfHeader: Various pseudopotential metadata

mesh::PseudoPotentialIO.UpfMesh: Radial mesh, mesh integration factors, and other mesh information

nlcc::Union{Nothing, Vector{Float64}}: Pseudized core charge on the radial grid, (ignored if core_correction is false)

local_::Union{Nothing, Vector{Float64}}: Local part of the pseudopotential on the radial grid (ignored if is_coulomb)

nonlocal::PseudoPotentialIO.UpfNonlocal: Nonlocal part of the pseudopotential

pswfc::Vector{PseudoPotentialIO.UpfChi}: Pseudo-atomic valence wavefunctions

full_wfc::Union{Nothing, PseudoPotentialIO.UpfFullWfc}: All-electron wavefunctions

rhoatom::Vector{Float64}: Pseudo-atomic valence charge density on the radial grid with prefactor 4πr²

spin_orb::Union{Nothing, PseudoPotentialIO.UpfSpinOrb}: Spin-orbit coupling data, (ignored if has_so is false)

paw::Union{Nothing, PseudoPotentialIO.UpfPaw}: PAW data, (ignored if is_paw is false)

gipaw::Union{Nothing, PseudoPotentialIO.UpfGipaw}: GIPAW data

PseudoPotentialIO.Psp8File — Type

struct Psp8File <: PsPFile

ABINIT PSeudoPotential format 8 file contents. Information on the file format specification and the meaning of the quantities within the file can be found on the "psp8" page of the ABINIT documentation.

identifier::String: Identifier

header::PseudoPotentialIO.Psp8Header: Various pseudopotential metadata

rgrid::Vector{Float64}: Uniform radial grid starting at r = 0.0

v_local::Vector{Float64}: Local part of the pseudopotential

projectors::Vector{Vector{Vector{Float64}}}: Radial part of the Kleinman-Bylander projectors for each angular momentum

ekb::Vector{Vector{Float64}}: Kleinman-Bylander energies for each angular momentum

projectors_so::Union{Nothing, Vector{Vector{Vector{Float64}}}}: Radial part of the spin-orbit Kleinman-Bylander projectors for each angular momentum

ekb_so::Union{Nothing, Vector{Vector{Float64}}}: Spin-orbit Kleinman-Bylander energies for each angular momentum

rhoc::Union{Nothing, Vector{Float64}}: Model core charge density with 4π prefactor

d_rhoc_dr::Union{Nothing, Vector{Float64}}: First derivative of the model core charge density

d2_rhoc_dr2::Union{Nothing, Vector{Float64}}: Second derivative of the model core charge density

d3_rhoc_dr3::Union{Nothing, Vector{Float64}}: Third derivative of the model core charge density

d4_rhoc_dr4::Union{Nothing, Vector{Float64}}: Fourth derivative of the model core charge density

rhov::Union{Nothing, Vector{Float64}}: Valence charge density with 4π prefactor

ae_rhov::Union{Nothing, Vector{Float64}}: All-electron valence charge density with 4π prefactor

ae_rhoc::Union{Nothing, Vector{Float64}}: All-electron core charge density with 4π prefactor

PseudoPotentialIO.HghFile — Type

struct HghFile <: PsPFile

Hartwigsen-Goedecker-Hutter pseudopotential file contents.

identifier::String: Identifier

functional::String: Functional guessed from the source file path if available

title::String: Description

zion::Vector{Int64}: Pseudo-atomic (valence) charge

rloc::Float64: Cutoff radius for the local part of the pseudopotential

nloc::Int64: Number of coefficients defining the local part of the pseudopotential

cloc::Vector{Float64}: Coefficients of the local part of the pseudopotential

lmax::Int64: Maximum angular momentum

rp::Vector{Float64}: Non-local projector cutoff radius for each angular momentum

h::Vector{Matrix{Float64}}: Kleinman-Bylander energies

Functions

Loading and saving

PseudoPotentialIO.load_psp_file — Function

load_psp_file(path::AbstractString) -> Any

Parse a pseudopotential file into a PsPFile struct.

load_psp_file(path)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/load.jl:8.

PseudoPotentialIO.save_psp_file — Function

save_psp_file(
    path::AbstractString,
    psp::PsPFile,
    file_format::String
)

Save a PsPFile struct into a file.

save_psp_file(path, psp, file_format)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/save.jl:4.

Conversion

PseudoPotentialIO.UpfFile — Method

UpfFile(file::Psp8File) -> UpfFile

UpfFile(file::Psp8File)

Convert a Psp8File into an equivalent UpfFile.

UpfFile(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/conversion/to_upf.jl:83.

Metadata

PseudoPotentialIO.identifier — Function

identifier(file::PsPFile) -> String

Identifying data.

identifier(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/file.jl:12.

identifier(psp)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/hgh.jl:102.

identifier(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/psp8.jl:354.

identifier(psp)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/upf.jl:419.

PseudoPotentialIO.format — Function

format(file::PsPFile) -> String

Pseudopotential file format.

format(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/file.jl:17.

format(_)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/hgh.jl:103.

format(_)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/psp8.jl:355.

format(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/upf.jl:420.

PseudoPotentialIO.element — Function

element(file::PsPFile) -> PeriodicTable.Element

The element which the pseudopotential was constructed to reproduce.

element(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/file.jl:22.

element(psp)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/hgh.jl:104.

element(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/psp8.jl:356.

element(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/upf.jl:452.

PseudoPotentialIO.functional — Function

functional(file::PsPFile) -> Vector{Symbol}

Exchange-correlation functional identifiers (Libxc-compatible), or empty if unknown.

functional(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/file.jl:27.

functional(psp)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/hgh.jl:111.

functional(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/psp8.jl:357.

functional(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/upf.jl:451.

PseudoPotentialIO.valence_charge — Function

valence_charge(file::PsPFile) -> Int64

Pseudo-atomic core charge.

valence_charge(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/file.jl:32.

valence_charge(psp)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/hgh.jl:112.

valence_charge(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/psp8.jl:358.

valence_charge(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/upf.jl:458.

PseudoPotentialIO.is_norm_conserving — Function

is_norm_conserving(file::PsPFile) -> Bool

Whether the pseudopotential is of the norm-conserving kind.

is_norm_conserving(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/file.jl:37.

is_norm_conserving(_)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/hgh.jl:113.

is_norm_conserving(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/psp8.jl:359.

is_norm_conserving(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/upf.jl:453.

PseudoPotentialIO.is_ultrasoft — Function

is_ultrasoft(file::PsPFile) -> Bool

Whether the pseudopotential is of the ultrasoft kind.

is_ultrasoft(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/file.jl:42.

is_ultrasoft(_)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/hgh.jl:114.

is_ultrasoft(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/psp8.jl:360.

is_ultrasoft(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/upf.jl:454.

PseudoPotentialIO.is_paw — Function

is_paw(file::PsPFile) -> Bool

Whether the pseudopotential is of the plane-augmented wave kind.

is_paw(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/file.jl:47.

is_paw(_)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/hgh.jl:115.

is_paw(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/psp8.jl:361.

is_paw(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/upf.jl:455.

PseudoPotentialIO.has_spin_orbit — Function

has_spin_orbit(file::PsPFile) -> Bool

Whether the pseudopotential contains relativistic spin-orbit coupling data.

has_spin_orbit(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/file.jl:52.

has_spin_orbit(_)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/hgh.jl:116.

has_spin_orbit(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/psp8.jl:362.

has_spin_orbit(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/upf.jl:456.

PseudoPotentialIO.has_model_core_charge_density — Function

has_model_core_charge_density(file::PsPFile) -> Bool

Whether the pseudopotential contains a model core charge density used in non-linear core correction.

has_model_core_charge_density(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/file.jl:57.

has_model_core_charge_density(_)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/hgh.jl:117.

has_model_core_charge_density(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/psp8.jl:363.

has_model_core_charge_density(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/upf.jl:457.

PseudoPotentialIO.formalism — Function

formalism(file::PsPFile) -> String

Formalism of the pseudopotential.

formalism(file)

defined at /home/runner/work/PseudoPotentialIO.jl/PseudoPotentialIO.jl/src/file/file.jl:62.